statically rendering slightly more of the header
